+++ .gitignore
@@ -6,3 +6,4 @@
 /build/
 /dist/
 /out/
+.env
